Add TestCase class in tests
